Ajustes presupueto

+++ AutomatMediciones.Libs/Dtos/ProductoPrecioDto.cs
@@ -4,8 +4,8 @@
     {
         public string RecID { get; set; }
         public string IDProducto { get; set; }
-        public double Precio { get; set; }
-        public double Costo { get; set; }
+        public decimal Precio { get; set; }
+        public decimal Costo { get; set; }
         public int NroMonedaPrecio { get; set; }
         public int NroLista { get; set; }
     }
